Writing state to disk and read from disk
