+++ .gitignore
@@ -1,6 +1,6 @@
 # Development
 test.db*
-e2e-test/container_etc/data.db
+e2e-test/container_etc/state.json
 
 # Compiled binaries
 bin/
