+++ e2e-test/.dockerignore
@@ -1,3 +1,4 @@
 node_modules/
 *.log
 results/
+container_etc/
