+++ go.mod
@@ -12,7 +12,7 @@ require (
 	github.com/Azure/go-autorest/autorest/validation v0.1.0 // indirect
 	github.com/gin-gonic/gin v1.4.0
 	github.com/gobuffalo/packr/v2 v2.5.2
-	github.com/gofrs/uuid v3.2.0+incompatible
+	github.com/google/renameio v0.1.0
 	github.com/jinzhu/copier v0.0.0-20190625015134-976e0346caa8
 	github.com/spf13/viper v1.4.0
 	golang.org/x/crypto v0.0.0-20190701094942-4def268fd1a4
