+++ utils/fs.go
@@ -1,0 +1,120 @@
+/*
+Copyright © 2019 Alessandro Segala (@ItalyPaleAle)
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU Affero General Public License as published
+by the Free Software Foundation, version 3 of the License.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU Affero General Public License for more details.
+
+You should have received a copy of the GNU Affero General Public License
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
+*/
+
+package utils
+
+import (
+	"io"
+	"os"
+	"path/filepath"
+)
+
+// RemoveContents remove all contents within a directory
+// Source: https://stackoverflow.com/a/33451503/192024
+func RemoveContents(dir string) error {
+	d, err := os.Open(dir)
+	if err != nil {
+		return err
+	}
+	defer d.Close()
+	names, err := d.Readdirnames(-1)
+	if err != nil {
+		return err
+	}
+	for _, name := range names {
+		err = os.RemoveAll(filepath.Join(dir, name))
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+// CopyFile copies the src file to dst. Any existing file will be overwritten and will not copy file attributes.
+// Source: https://stackoverflow.com/a/21061062/192024
+func CopyFile(src string, dst string) error {
+	in, err := os.Open(src)
+	if err != nil {
+		return err
+	}
+	defer in.Close()
+
+	out, err := os.Create(dst)
+	if err != nil {
+		return err
+	}
+	defer out.Close()
+
+	_, err = io.Copy(out, in)
+	if err != nil {
+		return err
+	}
+	return out.Close()
+}
+
+// PathExists returns true if the path exists on disk
+func PathExists(path string) (bool, error) {
+	_, err := os.Stat(path)
+	if os.IsNotExist(err) {
+		return false, nil
+	}
+	return err == nil, err
+}
+
+// FolderExists returns true if the path exists on disk and it's a folder
+func FolderExists(path string) (bool, error) {
+	info, err := os.Stat(path)
+	if err != nil {
+		// Ignore the error if it's a "not exists", that's the goal
+		if os.IsNotExist(err) {
+			err = nil
+		}
+		return false, err
+	}
+	if info.IsDir() {
+		// Exists and it's a folder
+		return true, nil
+	}
+	// Exists, but not a folder
+	return false, nil
+}
+
+// EnsureFolder creates a folder if it doesn't exist already
+func EnsureFolder(path string) error {
+	exists, err := PathExists(path)
+	if err != nil {
+		return err
+	} else if !exists {
+		if err := os.Mkdir(path, os.ModePerm); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+// SymlinkAtomic creates a symbolic link atomically, replacing existing ones if any
+func SymlinkAtomic(src string, dest string) error {
+	// First, create a temporary symlink, then rename it, so the entire operation is atomic
+	tmpLink := dest + "_" + RandString(6)
+	if err := os.Symlink(src, tmpLink); err != nil {
+		return err
+	}
+	if err := os.Rename(tmpLink, dest); err != nil {
+		return err
+	}
+	return nil
+}
